feat: implement ResourceCard and SelectionToolbar components for enhanced resource management
